+++ src/Application/Countdown/UseCase/DeleteCountdownUseCase.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace App\Application\Countdown\UseCase;
 
 use App\Domain\Countdown\CountdownException;
+use App\Domain\Event\EventPublisher;
 use App\Infrastructure\Helper\CountdownValidationHelper;
 use App\Infrastructure\Persistence\PDOCountdownRepository;
 use Psr\Log\LoggerInterface;
@@ -20,6 +21,7 @@ readonly class DeleteCountdownUseCase
      */
     public function __construct(
         private PDOCountdownRepository    $repository,
+        private EventPublisher            $eventPublisher,
         private LoggerInterface           $logger,
         private CountdownValidationHelper $validator,
     ){}
@@ -39,11 +41,21 @@ readonly class DeleteCountdownUseCase
 
         $this->validator->validateId($id);
 
+        $countdown = $this->repository->findById($id);
+        if ($countdown === null) {
+            throw CountdownException::notFound($id);
+        }
+
         $result = $this->repository->delete($id);
 
         if (!$result) {
             throw CountdownException::failedToDelete();
         }
+
+        $countdown->markDeleted();
+        $events = $countdown->getDomainEvents();
+        $this->eventPublisher->publishAll($events);
+        $countdown->clearDomainEvents();
 
         $this->logger->info('Countdown delete finished',
             [
